requirements.txt
python-dateutil==2.8.2
regex==2022.9.13
ttkbootstrap==1.9.0